requirements.txt
pytest==8.4.2
selenium==4.36.0
python-dotenv==1.1.1
pytest-md-report==0.7.0
pytest-playwright==0.7.1
playwright>=1.55.0
allure-pytest==2.15.0